fix gitpod

+++ scripts/setup.sh
@@ -1,7 +1,9 @@
 #!/usr/bin/env bash
 npm install -g npm@latest node-red@latest svelte-integration-red@latest 
 # Initialize sample project
-rm -r /home/gitpod/.node-red
+if [ -d /home/gitpod/.node-red ]; then
+    rm -r /home/gitpod/.node-red
+fi
 mkdir /home/gitpod/.node-red
 cp /workspace/node-red-contrib-oauth2/example-project/* /home/gitpod/.node-red/
 cd /home/gitpod/.node-red/
